requirements.txt
pydantic==2.13.3
instructor==1.15.1
rapidfuzz==3.13.0
openai==2.32.0
anthropic==0.97.0
python-dotenv==1.2.1
pandas==2.3.3
tqdm==4.67.3
mapie==1.3.0
scikit-learn==1.8.0
numpy==2.4.4
PyYAML==6.0.3
